Properly restore irregular ROIs.
If the pre-existing ROI does not have a rectangular shape, the
getROI/setROI methods of ImagePlus are not enough. We also need
to save a binary mask representing the irregularly shaped ROI
and reapply that mask.

The binary mask only has the size of the rectangular bounding box
surrounding the ROI, so it cannot be used alone, it must be
applied after restoring the rectangular ROI.