Add license.
My understanding of the QMUL policy in intellectual property is that
software developed by a QMUL employee in the course of their employment,
but which may not be reasonably foreseen to have commercial potential,
is not treated as an Invention (which would belong to QMUL) but as an
Academic Work whose property remains with the employee.

I don't believe this software can reasonably be foreseen to have any
commercial potential. As such, it is thus my property and therefore I
am free to release it under the GNU General Public License.