Use drop-down menus for script parameters when possible.
When a script expects a thresholding algorithm as parameter, use a
drop-down menu so that the user can select one of the available
algorithm instead of having to manually enter its name.

Unfortunately the list of algorithms has to be explicitly specified in
each script, it cannot be automatically generated.